pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "openwebui-chat-client"
version = "0.1.23"
authors = [
    { name="fujie", email="fj1945@live.cn" },
]
description = "A powerful and intelligent Python client for the Open WebUI API."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "GPL-3.0-only" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "requests",
    "python-dotenv",
    "httpx"
]

[project.optional-dependencies]
test = [
    "responses"
]

[project.urls]
Homepage = "https://github.com/Fu-Jie/openwebui-chat-client"
"Bug Tracker" = "https://github.com/Fu-Jie/openwebui-chat-client/issues"
"Documentation" = "https://github.com/Fu-Jie/openwebui-chat-client#readme"

[tool.setuptools.packages.find]
include = ["openwebui_chat_client*"]
